cleanup verbose message handling

+++ icmake/icm-pre/process/punlinkgch.c
@@ -4,8 +4,10 @@ void pUnlinkGch(int idx)
 {
     char const *path = depGch(idx);
 
-    if (sproc.d_go == GO)
-        unlink(path);
+    if (!rss_exists(path))
+        return;
 
     optMsg(1, "unlink(%s)", path);
+    if (sproc.d_go == GO)
+        unlink(path);
 }
